add old emu.py back
